refactor: rename influx_service.py to service.py and influx_base.py to base.py for consistency with other modules
