+++ sigenergy2mqtt/influxdb/base.py
